full dataset

+++ src/main/java/hello/QueryNumSpeedsOver100.java
@@ -83,11 +83,11 @@ public class QueryNumSpeedsOver100 {
             //Result resultTotalSpeed = table.get(theGetTotalSpeed);
             //int totalSpeed = Integer.parseInt(Bytes.toString(resultTotalSpeed.getValue(Bytes.toBytes("results"), Bytes.toBytes("count"))));
 
-            double totalSpeed = 0;
+            long totalSpeed = 0;
             for (int i = 0; i <= 5; i++){
                 Get theGetTotalSpeed = new Get(Bytes.toBytes("Total Speed "+ i + ":"));
                 Result resultTotalSpeed = table.get(theGetTotalSpeed);
-                totalSpeed += Double.parseDouble(Bytes.toString(resultTotalSpeed.getValue(Bytes.toBytes("results"), Bytes.toBytes("count"))));
+                totalSpeed += Long.parseLong(Bytes.toString(resultTotalSpeed.getValue(Bytes.toBytes("results"), Bytes.toBytes("count"))));
             }
 
             Get theGetTotalRecords = new Get(Bytes.toBytes("Total speed records"));
